Prioritize customized tasks over predefined ones
To make sure that predefined tasks customized in the workspace's
'tasks.json' file are picked-up first when calling 'findTaskByName'.

Also make sure to prepend 'ada: ' to the GNAT SAS basic task names
in the GNAT SAS compound tasks.

Add a basic test for this.

For eng/ide/ada_language_server#1384