Handle partially qualified dotted-name completion
When completing an invisible item with a partially qualified prefix
(e.g: 'P2' instead of 'P1.P2.'), we should not try
to rely on the prefix being typed by the user: we should
just import the fully qualified unit.

Add a test for this.

For eng/ide/ada_language_server#1610