Add initial GNATcheck integration
This is done for now by having the Task Provider offer two new
tasks, one for running GNATcheck on the current file, and one
for running GNATcheck on the current project.

Augment the error pattern to recognize GNATcheck rule violations
as warnings.

Add mention of GNATcheck in the documentation for the extension.
Regenerate the screenshot showing the tasks, to show the GNATcheck
tasks.

For eng/projects/scheme/scheme-deliverables#86